flask/requirements.txt
click==8.0.3
Flask==2.0.2
Flask_WTF
wtforms-widgets
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug==2.0.2
